Move some bits around
